сделал 4 лабу
